requirements.txt
Flask==3.1.0
gunicorn==23.0.0
pandas==2.2.3
numpy
flask-socketio
google-cloud-firestore
eventlet